requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy>=2.0.43
alembic==1.13.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
httpx==0.27.0
python-multipart==0.0.9
taxteclib>=0.3.2
openpyxl>=3.1.2
azure-devops>=7.1.0b4
msrest>=0.7.1

# Testing
pytest==8.2.0
pytest-cov==5.0.0
pytest-asyncio>=1.2.0
faker==25.0.0
robotframework==7.0